refactor: adopt underscore free api
