+++ tradingbot/ui/app.py
@@ -4,8 +4,8 @@ from __future__ import annotations
 
 from fastapi import FastAPI, HTTPException
 
-from tradingbot.core.runtime_controller import RuntimeController
-from tradingbot.core.validation_manager import ValidationManager
+from tradingbot.core.runtimecontroller import RuntimeController
+from tradingbot.core.validationmanager import ValidationManager
 
 
 runtime = RuntimeController()
@@ -17,25 +17,25 @@ def create_app() -> FastAPI:
 
     @app.get("/status")
     def status():
-        return runtime.get_state()
+        return runtime.getstate()
 
     @app.post("/live/{asset}/enable")
     def enable(asset: str):
         try:
-            runtime.enable_live(asset)
+            runtime.enablelive(asset)
         except ValueError as exc:
             raise HTTPException(status_code=400, detail=str(exc)) from exc
         return {"asset": asset, "live": True}
 
     @app.post("/live/{asset}/disable")
     def disable(asset: str):
-        runtime.disable_live(asset)
+        runtime.disablelive(asset)
         return {"asset": asset, "live": False}
 
     @app.post("/kill/global/{onoff}")
     def kill(onoff: str):
-        runtime.set_global_kill(onoff.lower() == "on")
-        return {"kill_switch": runtime.get_state()["global"]["kill_switch"]}
+        runtime.setglobalkill(onoff.lower() == "on")
+        return {"kill_switch": runtime.getstate()["global"]["kill_switch"]}
 
     @app.get("/validation/{strategy_id}")
     def validation(strategy_id: str):
